RETAKE Damageable Retake Projectile/AProjectileLauncher

+++ Unity/TowerDefenseWorkshop_Template/Assets/TD/Scripts/DamageableDetector.cs
@@ -86,7 +86,7 @@
 			}
 		}
 
-		private void Damageable_OnDamageTaken(Damageable caller, int currentHealth, int damageTaken)
+		private void Damageable_OnDamageTaken(Damageable caller, float currentHealth, float damageTaken)
 		{
 			if (currentHealth <= 0)
 			{
